+++ com.unity.cinemachine/Editor/PropertyDrawers/InputAxisPropertyDrawer.cs
@@ -7,15 +7,13 @@ namespace Unity.Cinemachine.Editor
     [CustomPropertyDrawer(typeof(InputAxis))]
     partial class InputAxisPropertyDrawer : PropertyDrawer
     {
-        InputAxis def = new (); // to access name strings
-
         public override VisualElement CreatePropertyGUI(SerializedProperty property)
         {
             // When foldout is closed, we display the axis value on the same line, for convenience
             var foldout = new Foldout { text = property.displayName, tooltip = property.tooltip };
             foldout.BindProperty(property);
 
-            var valueProp = property.FindPropertyRelative(() => def.Value);
+            var valueProp = property.FindPropertyRelative(nameof(InputAxis.Value));
             var valueLabel = new Label(" ") { style = { minWidth = InspectorUtility.SingleLineHeight * 2 }};
             var valueField =  new PropertyField(valueProp, "") { style = { flexGrow = 1, marginLeft = 2 }};
             valueField.OnInitialGeometry(() => valueField.SafeSetIsDelayed());
@@ -27,17 +25,30 @@ namespace Unity.Cinemachine.Editor
             foldout.AddChild(InspectorUtility.PropertyRow(valueProp, out var valueField2));
             valueField2.OnInitialGeometry(() => valueField2.SafeSetIsDelayed());
 
-            var centerField = foldout.AddChild(new PropertyField(property.FindPropertyRelative(() => def.Center)));
+            var centerProp = property.FindPropertyRelative(nameof(InputAxis.Center));
+            var centerRow = foldout.AddChild(InspectorUtility.PropertyRow(centerProp, out var centerField));
+            centerField.OnInitialGeometry(() => centerField.SafeSetIsDelayed());
+            centerRow.Contents.Add(new Button(() =>
+            {
+                valueProp.floatValue = centerProp.floatValue;
+                valueProp.serializedObject.ApplyModifiedProperties();
+            })
+            { 
+                text = "Recenter", 
+                tooltip = "Reset the axis value to the center",
+                style = { marginLeft = 4, marginRight = 0, marginTop = 0, marginBottom = 0 }
+            });
+
             var rangeContainer = foldout.AddChild(new VisualElement() { style = { flexDirection = FlexDirection.Row }});
-            rangeContainer.Add(new PropertyField(property.FindPropertyRelative(() => def.Range)) { style = { flexGrow = 1 }});
-            var wrapProp = property.FindPropertyRelative(() => def.Wrap);
+            rangeContainer.Add(new PropertyField(property.FindPropertyRelative(nameof(InputAxis.Range))) { style = { flexGrow = 1 }});
+            var wrapProp = property.FindPropertyRelative(nameof(InputAxis.Wrap));
             var wrap = rangeContainer.AddChild(new PropertyField(wrapProp, "")
                 { style = { alignSelf = Align.Center, marginLeft = 5, marginRight = 5, marginTop = 2 }});
             var wrapLabel = rangeContainer.AddChild(new Label(wrapProp.displayName)
                 { tooltip = wrapProp.tooltip, style = { alignSelf = Align.Center }});
-            var recentering = foldout.AddChild(new PropertyField(property.FindPropertyRelative(() => def.Recentering)));
+            var recentering = foldout.AddChild(new PropertyField(property.FindPropertyRelative(nameof(InputAxis.Recentering))));
 
-            var flagsProp = property.FindPropertyRelative(() => def.Restrictions);
+            var flagsProp = property.FindPropertyRelative(nameof(InputAxis.Restrictions));
             ux.TrackPropertyWithInitialCallback(flagsProp, (prop) =>
             {
                 if (prop.serializedObject == null)